Implement smart update vacancies algorithm. Improve package structure quality.
